+++ src/indexClient.js
@@ -1,4 +1,3 @@
-var _ = require( "lodash" );
 var indexClient = require( "nonstop-index-client" );
 
 function getClient( config ) {
@@ -16,8 +15,8 @@ function updateConfig( config ) {
 module.exports = function( config ) {
 	var state = {
 		config: config,
-		client: getClient( config ),
-		update: updateConfig.bind( this )
+		client: getClient( config )
 	};
+	state.update = updateConfig.bind( state );
 	return state;
 };